pyproject.toml
[tool.poetry]
name = "drf-homework"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
pillow = "^11.1.0"
ipython = "^8.32.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.4.0"
coverage = "^7.6.11"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"


[tool.poetry.group.lint.dependencies]
mypy = "^1.14.1"
black = "^25.1.0"
isort = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119
